fiture update profile

+++ routes/web.php
@@ -1,6 +1,7 @@
 <?php
 
 use Illuminate\Support\Facades\Route;
+use Illuminate\Support\Facades\Auth;
 
 /*
 |--------------------------------------------------------------------------
@@ -20,6 +21,8 @@ Route::get('/', function () {
 Auth::routes();
 
 Route::get('/home', 'HomeController@index')->name('home');
+Route::get('/profile', 'Admin\ProfileController@edit')->name('profile.edit');
+Route::patch('profile', 'Admin\ProfileController@update')->name('profile.update');
 
 Route::group(['middleware'=> ['auth','admin']], function(){
 
@@ -27,11 +30,12 @@ Route::group(['middleware'=> ['auth','admin']], function(){
     return view('admin.dashboard');
     });
 
-    Route::get('/role-register','Admin\DashboardController@registered');
+    Route::get('/role-register','Admin\DashboardController@registered')->name('role-register');
 
-    Route::get('/role-edit/{id}','Admin\DashboardController@registeredit');
-    Route::put('/role-register-update/{id}','Admin\DashboardController@registerupdate');
-    Route::delete('/role-delete/{id}','Admin\DashboardController@registerdelete');
+    Route::get('/role-edit/{id}','Admin\DashboardController@registeredit')->name('role-edit');
+    Route::put('/role-register-update/{id}','Admin\DashboardController@registerupdate')->name('role-register');
+    Route::delete('/role-delete/{id}','Admin\DashboardController@registerdelete')->name('role-delete');
 
-    Route::get('/abouts','Admin\AboutusController@index');
+    Route::get('/abouts','Admin\AboutusController@index')->name('abouts');
+    
 });
